Send local_id with remix API call
